+++ src/plugins/ncr/components/node-picker-field/index.jsx
@@ -22,6 +22,7 @@ class NodePickerField extends React.Component {
     handleGetAll: PropTypes.func.isRequired,
     handleSearch: PropTypes.func.isRequired,
     isEditMode: PropTypes.bool,
+    isFulfilled: PropTypes.bool.isRequired,
     isGetAll: PropTypes.bool,
     isMulti: PropTypes.bool,
     response: PropTypes.instanceOf(Message),
@@ -45,6 +46,7 @@ class NodePickerField extends React.Component {
     this.state = {
       hasLoadedFirstSet: false,
       existingNodes: [],
+      menuListScrollTop: 0,
     };
 
     this.handleLoadMore = debounce(this.handleLoadMore.bind(this), 500, {
@@ -56,13 +58,18 @@ class NodePickerField extends React.Component {
     this.handleMenuOpen = this.handleMenuOpen.bind(this);
     this.handlePick = this.handlePick.bind(this);
     this.handleRemove = this.handleRemove.bind(this);
+    this.handleScroll = debounce(this.handleScroll.bind(this), 500, {
+      leading: false,
+      trailing: true,
+    });
   }
 
   /**
    * getting new nodes resets the scroll position, correct that here
    */
   componentDidUpdate(prevProps) {
-    const { constants, response } = this.props;
+    const { response } = this.props;
+    const { menuListScrollTop } = this.state;
     if (!prevProps.response || !response) {
       return;
     }
@@ -71,9 +78,7 @@ class NodePickerField extends React.Component {
       && response.get('ctx_request').get('page') > 1
       && !response.get('response_id').equals(prevProps.response.get('response_id'))
     ) {
-      this.menuList.scrollTop = Array.from(this.menuList.children)
-        .slice(0, prevProps.response.get('ctx_request').get('page') * constants.REQUEST_COUNT)
-        .reduce((acc, { offsetHeight }) => (acc + offsetHeight), 0);
+      this.menuList.scrollTop = menuListScrollTop;
     }
   }
 
@@ -119,6 +124,7 @@ class NodePickerField extends React.Component {
         ...existingNodes,
         ...props.response.get('nodes', []),
       ],
+      menuListScrollTop: this.menuList.scrollTop,
     }), () => {
       handleSearch(response.get('ctx_request').get('q'), response.get('ctx_request').get('page') + 1);
     });
@@ -149,6 +155,13 @@ class NodePickerField extends React.Component {
     fields.remove(fields.getAll().findIndex((ref) => ref.equals(nodeRef)));
   }
 
+  handleScroll(isScrolledToBottom) {
+    const { isFulfilled } = this.props;
+    if (isScrolledToBottom && isFulfilled) {
+      this.handleLoadMore();
+    }
+  }
+
   handleSelect(nodeRef) {
     const { fields } = this.props;
     const isSelected = !!(fields.getAll() || []).find((ref) => ref.equals(nodeRef));
@@ -174,6 +187,10 @@ class NodePickerField extends React.Component {
               {...props}
               {...this.props}
               {...this.state}
+              innerProps={{
+                ...props.innerProps,
+                onScroll: ({ target }) => this.handleScroll(target.scrollHeight - target.scrollTop === target.clientHeight),
+              }}
             >
               {props.children}
             </MenuComponent>
@@ -204,7 +221,6 @@ class NodePickerField extends React.Component {
         onChange={this.handleChange}
         onInputChange={this.handleInputChange}
         onMenuOpen={this.handleMenuOpen}
-        onMenuScrollToBottom={this.handleLoadMore}
         options={existingNodes.concat(!response ? [] : response.get('nodes', [])).map((node) => ({
           label: node.get('title'),
           value: NodeRef.fromNode(node),